Move FakeFileHTTPServerBasedTestCase to tests
